Organized ML model coupling and temp trend examples

+++ eo_stack.py
@@ -171,26 +171,31 @@ class opticalTempStack(eoTempStack):
         calc_band = []
         locations = []
         
-        for date in self.getTempData('nir'):
-            nir = self.getBand('nir', date=date)
-            # Allow division by zero
-            np.seterr(divide='ignore', invalid='ignore')
-            # Calculate index
-            if index == 'NDVI': #(nir-red)/(nir+red)
-                red = self.getBand('red', date=date)
-                calc_band.append((nir.astype(float) - red.astype(float)) / (nir + red))
-            if index == 'LSWI': #(nir-swir1)/(nir+swir1)
-                swir1 = self.getBand('swir1', date=date)
-                calc_band.append((nir.astype(float) - swir1.astype(float)) / (nir + swir1))
-                
-        for id, date in enumerate(self.getBandsLoc('nir')):
+        ## Iterate to read bands and calculate (first check if file exists)
+        for id, date in enumerate(self.getTempData('nir')):
+            # define location of index raster file
             bandLocation = self.out_directory+index+'_'+self.prod_type+'_'+self.bands_temporal_range['nir'][id].strftime('%Y%m%d')+'_.tif'
-            locations.append(bandLocation)
-            with rasterio.open(date) as src:
-                kwargs = src.meta
-            kwargs.update(dtype=rasterio.float32)
-            with rasterio.open(bandLocation, 'w', **kwargs) as dst:
-                dst.write_band(1, calc_band[id].astype(rasterio.float32))
+            # check if raster file already exists
+            if os.path.isfile(bandLocation):
+                locations.append(bandLocation)
+            # if not, calculate index and put in raster file
+            else:
+                nir = self.getBand('nir', date=date)
+                # Allow division by zero
+                np.seterr(divide='ignore', invalid='ignore')
+                # Calculate index
+                if index == 'NDVI': #(nir-red)/(nir+red)
+                    red = self.getBand('red', date=date)
+                    calc_band.append((nir.astype(float) - red.astype(float)) / (nir + red))
+                if index == 'LSWI': #(nir-swir1)/(nir+swir1)
+                    swir1 = self.getBand('swir1', date=date)
+                    calc_band.append((nir.astype(float) - swir1.astype(float)) / (nir + swir1))
+                locations.append(bandLocation)
+                with rasterio.open(self.getBandsLoc('nir')[id]) as src:
+                    kwargs = src.meta
+                kwargs.update(dtype=rasterio.float32)
+                with rasterio.open(bandLocation, 'w', **kwargs) as dst:
+                    dst.write_band(1, calc_band[id].astype(rasterio.float32))
         
         # Update object variables
         self.setBandsLoc(index, locations)
@@ -279,8 +284,15 @@ class S2TempStack(opticalTempStack):
         else:
             temp_range = {}
             for key, value in self.getBandsLoc().iteritems():
-                temp_range[key] = list(map(lambda x: datetime.datetime.strptime(x.split('/')[-2][11:19], 
+                try:
+                    temp_range[key] = list(map(lambda x: datetime.datetime.strptime(x.split('/')[-2][11:19], 
                                                                                 '%Y%m%d').date(),value))
+                except:
+                    try:
+                        temp_range[key] = list(map(lambda x: datetime.datetime.strptime(x.split('/')[-2][25:33], 
+                                                                                '%Y%m%d').date(),value))
+                    except ValueError:
+                        raise Exception('The S2 L2A product does not follow naming conventions (dates %Y%m%d at 11:19 or 25:33).')
             self.bands_temporal_range = temp_range
